+++ src/cli/fmt.rs
@@ -23,7 +23,7 @@ impl Fmt {
 
     pub fn log_n_plus_one(show_npo: bool, config: &Config) {
         let n_plus_one_info = config.n_plus_one();
-        let mut message = format!("N + 1 detected: {}", n_plus_one_info.len());
+        let mut message = format!("N + 1 detected: {}", n_plus_one_info.map.len());
 
         if show_npo {
             message.push('\n');
@@ -31,175 +31,5 @@ impl Fmt {
         }
 
         tracing::info!("{}", message);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::cli::fmt::Fmt;
-    use crate::core::config::{Config, Field, Http, Type};
-    #[test]
-    fn test_nplusone_resolvers() {
-        let config = Config::default().query("Query").types(vec![
-            (
-                "Query",
-                Type::default().fields(vec![(
-                    "f1",
-                    Field::default()
-                        .type_of("F1".to_string())
-                        .into_list()
-                        .http(Http::default()),
-                )]),
-            ),
-            (
-                "F1",
-                Type::default().fields(vec![(
-                    "f2",
-                    Field::default()
-                        .type_of("F2".to_string())
-                        .into_list()
-                        .http(Http::default()),
-                )]),
-            ),
-            (
-                "F2",
-                Type::default()
-                    .fields(vec![("f3", Field::default().type_of("String".to_string()))]),
-            ),
-        ]);
-        let actual = config.n_plus_one();
-        let formatted = Fmt::format_n_plus_one_queries(actual);
-        let mut formatted = formatted.split('\n').collect::<Vec<_>>();
-        formatted.sort();
-        insta::assert_snapshot!(formatted.join("\n"));
-    }
-
-    #[test]
-    fn test_nplusone_nested_resolvers() {
-        let config = Config::default().query("Query").types(vec![
-            (
-                "Query",
-                Type::default().fields(vec![(
-                    "f1",
-                    Field::default()
-                        .type_of("F1".to_string())
-                        .into_list()
-                        .http(Http::default()),
-                )]),
-            ),
-            (
-                "F1",
-                Type::default().fields(vec![(
-                    "f2",
-                    Field::default().type_of("F2".to_string()).into_list(),
-                )]),
-            ),
-            (
-                "F2",
-                Type::default().fields(vec![(
-                    "f3",
-                    Field::default().type_of("F3".to_string()).into_list(),
-                )]),
-            ),
-            (
-                "F3",
-                Type::default().fields(vec![(
-                    "f4",
-                    Field::default()
-                        .type_of("String".to_string())
-                        .http(Http::default()),
-                )]),
-            ),
-        ]);
-
-        let actual = config.n_plus_one();
-        let formatted = Fmt::format_n_plus_one_queries(actual);
-        let mut formatted = formatted.split('\n').collect::<Vec<_>>();
-        formatted.sort();
-        insta::assert_snapshot!(formatted.join("\n"));
-    }
-
-    #[test]
-    fn test_nplusone_nested_resolvers_non_list_resolvers() {
-        let config = Config::default().query("Query").types(vec![
-            (
-                "Query",
-                Type::default().fields(vec![(
-                    "f1",
-                    Field::default()
-                        .type_of("F1".to_string())
-                        .http(Http::default()),
-                )]),
-            ),
-            (
-                "F1",
-                Type::default().fields(vec![(
-                    "f2",
-                    Field::default().type_of("F2".to_string()).into_list(),
-                )]),
-            ),
-            (
-                "F2",
-                Type::default().fields(vec![(
-                    "f3",
-                    Field::default().type_of("F3".to_string()).into_list(),
-                )]),
-            ),
-            (
-                "F3",
-                Type::default().fields(vec![(
-                    "f4",
-                    Field::default()
-                        .type_of("String".to_string())
-                        .http(Http::default()),
-                )]),
-            ),
-        ]);
-
-        let actual = config.n_plus_one();
-        let formatted = Fmt::format_n_plus_one_queries(actual);
-        let mut formatted = formatted.split('\n').collect::<Vec<_>>();
-        formatted.sort();
-        insta::assert_snapshot!(formatted.join("\n"));
-    }
-
-    #[test]
-    fn test_nplusone_cycles_with_resolvers() {
-        let config = Config::default().query("Query").types(vec![
-            (
-                "Query",
-                Type::default().fields(vec![(
-                    "f1",
-                    Field::default()
-                        .type_of("F1".to_string())
-                        .into_list()
-                        .http(Http::default()),
-                )]),
-            ),
-            (
-                "F1",
-                Type::default().fields(vec![
-                    ("f1", Field::default().type_of("F1".to_string()).into_list()),
-                    (
-                        "f2",
-                        Field::default()
-                            .type_of("String".to_string())
-                            .http(Http::default()),
-                    ),
-                ]),
-            ),
-            (
-                "F2",
-                Type::default()
-                    .fields(vec![("f3", Field::default().type_of("String".to_string()))]),
-            ),
-        ]);
-
-        let actual = config.n_plus_one();
-
-        let formatted = Fmt::format_n_plus_one_queries(actual);
-        let mut formatted = formatted.split('\n').collect::<Vec<_>>();
-        formatted.sort();
-        insta::assert_snapshot!(formatted.join("\n"));
     }
 }
