+++ src/core/config/npo/mod.rs
@@ -1,4 +1,6 @@
 mod display;
 mod n_plus_one;
+mod r#yield;
 
 pub use n_plus_one::*;
+pub use r#yield::*;
